Fixed several warnings and TEST Makefile (Itraq test build errors)
git-svn-id: https://open-ms.svn.sourceforge.net/svnroot/open-ms/OpenMS@3927 6adb6e08-d915-0410-941f-83917bcadc18

+++ source/APPLICATIONS/TOPP/SILACAnalyzer.C
@@ -438,7 +438,7 @@ class TOPPSILACAnalyzer
 	  				best_n = i;
 	  			}			
 	  		}
-	  		best_n = cluster_number_scaling * best_n; // slightly increase cluster number
+	  		best_n = UInt(cluster_number_scaling * best_n); // slightly increase cluster number
 	  		cuttree(size, tree, best_n, clusterid);
 	  		cluster_number[charge] = best_n;
